Fixed compiler warning on windows.
git-svn-id: https://svn.opendtect.org/od/trunk@27402 a985855d-d384-4b45-ae33-31a410998c3c

+++ plugins/uiHorizonAttrib/uiflattenedcube.cc
@@ -64,7 +64,7 @@ uiWriteFlattenedCube::uiWriteFlattenedCube( uiParent* p, EM::ObjectID horid )
     pp_.getZRange( horzrg_ );
     MouseCursorManager::restoreOverride();
     defzval_ = horzrg_.center() * SI().zDomain().userFactor();
-    defzval_ = mNINT32(defzval_);
+    defzval_ = (float) mNINT32(defzval_);
     zvalfld_ = new uiGenInput( this, txt, FloatInpSpec(defzval_) );
     zvalfld_->attach( alignedBelow, seisselin_ );
